Add time picker props table
