+++ docs/src/Routes/Demos/TimePickerDemo.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypesTable from '_shared/PropTypesTable';
 import SourcablePanel from '_shared/SourcablePanel';
 import { Typography } from 'material-ui';
 
@@ -29,6 +30,8 @@ const DatePickerDemoDemo = () => (
       title="Keyboard input"
       sourceFile="Demo/KeyboardTimePicker.jsx"
     />
+
+    <PropTypesTable src="TimePicker/TimePickerWrapper.jsx" />
   </div>
 );
